+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added common/NiPart as `TmPart` @okwme
 - Added common/NiTabBar as `TmTabBar` @okwme
 - Added common/NiToolBar as `TmToolBar` @okwme
+- Added `TmCookieConsent` @nylira
 
 ### Changed